+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 <!--Releasenotes start-->
 - The backend database now uses a more efficient method of storing video ID's.
-- Fixed a bug where the newest videos would not be fetched in a certain case.
+- Fixed a bug where the newest videos would not be fetched from the YouTube API in a certain case.
 <!--Releasenotes end-->
 
 ## v0.2.2